+++ src/MainTransformer/MainTransformer.php
@@ -14,7 +14,8 @@ declare(strict_types=1);
 namespace Rekalogika\Mapper\MainTransformer;
 
 use Rekalogika\Mapper\Exception\LogicException;
-use Rekalogika\Mapper\Exception\UnableToFindSuitableTransformerException;
+use Rekalogika\Mapper\MainTransformer\Exception\CannotFindTransformerException;
+use Rekalogika\Mapper\MainTransformer\Exception\TransformerReturnsUnexpectedValueException;
 use Rekalogika\Mapper\ObjectCache\ObjectCache;
 use Rekalogika\Mapper\ObjectCache\ObjectCacheFactoryInterface;
 use Rekalogika\Mapper\Transformer\Contracts\MainTransformerAwareInterface;
@@ -22,6 +23,7 @@ use Rekalogika\Mapper\Transformer\Contracts\MixedType;
 use Rekalogika\Mapper\Transformer\Contracts\TransformerInterface;
 use Rekalogika\Mapper\TransformerRegistry\TransformerRegistryInterface;
 use Rekalogika\Mapper\TypeResolver\TypeResolverInterface;
+use Rekalogika\Mapper\Util\TypeCheck;
 use Symfony\Component\PropertyInfo\Type;
 
 class MainTransformer implements MainTransformerInterface
@@ -150,9 +152,13 @@ class MainTransformer implements MainTransformerInterface
                 context: $context
             );
 
+            if (!TypeCheck::isVariableInstanceOf($result, $targetType)) {
+                throw new TransformerReturnsUnexpectedValueException($targetType, $result);
+            }
+
             return $result;
         }
 
-        throw new UnableToFindSuitableTransformerException($sourceTypes, $targetTypes);
+        throw new CannotFindTransformerException($sourceTypes, $targetTypes);
     }
 }