+++ src/Transformer/Exception/NotMappableValueException.php
@@ -11,8 +11,10 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Exception;
+namespace Rekalogika\Mapper\Transformer\Exception;
 
-abstract class MissingMemberTypeException extends NotMappableValueException
+use Rekalogika\Mapper\Exception\UnexpectedValueException;
+
+abstract class NotMappableValueException extends UnexpectedValueException
 {
 }